remove TuiDrawMode
remove because upon working on backend implementation, it became apparent this is not useful as intended. Might allow for PBO support in the future to increase pixel transfer speed, but not for now.
remove associated functions
remove as member of TuiTexture and from TuiTexture create function parameters
remove TUI_ERROR_INVALID_DRAW_MODE and associated strings and functions.