remove tuiWindowSizeFunction
glfwGetWindowSize and glfwWindowSizeFunction callbacks are not consistent across all platforms. It says in the GLFW docs that on some platforms, the size for these will be the entire window size including the title bar, while on others it will only be about the framebuffer area. Meanwhile, the simillar glfw functions with "framebuffer size" in their name are consistently across all platforms refering to the size of the render area of the window. Due to the inconsistency, I removed all uses of the "Size" versions and left only the "Framebuffer Size" versions.