finish multi window/ single window systems
A GLFWwindow is created on system initialization and stored in TuiSystem as BaseWindow.
if MultiWindow mode is disabled, the first window created adopts the BaseWindow.
if MultiWIndow mode is disabled, when the window is destroyed the BaseWindow is hidden
if MultiWindow mode is enabled, the BaseWindow is used as the base context that all  TuiWindow windows share from