remember window position and keep track of fullscreen
glfwGetWindowMonitor sometimes returns NULL for fullscreen windows, because GLFW sometimes makes fullscreen windows borderless. Add a TuiBoolean member to TuiWindow to keep track of if it is fullscreen instead.
Remove window position arguments from window set windowed functions. Instead, remember the position from where the window was when it was last windowed and use that instead.