seperate TuiBatch_s into other structs for each layout flag
The first property is always the detail mode. C and C++ standards ensure that the first property of a POD struct is always at the same memory address as the struct itself. The detail mode can be used to determine the struct's layout flag to then caste it into its appropriate type.
add a struct type with the properties that will be used for the new TUI_DETAIL_FLAG_LAYOUT_FREE.
all functions that use TuiBatch objects will need to be modified to use this new system